backend/requirements.txt
fastapi>=0.111.0
bcrypt>=4.0.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
greenlet>=3.0.0
alembic>=1.13.0
pydantic-settings>=2.2.0
httpx>=0.27.0
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
structlog>=24.1.0
aiosqlite>=0.20.0
aiofiles>=23.0.0
websockets>=12.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio>=4.3.0
httpx[http2]>=0.27.0
